+++ check_log.py
@@ -356,7 +356,7 @@ def main():
 
     file_chk_list = ["-f", "-i", "-m", "-o", "-F"]
     file_crt_list = ["-m", "-o"]
-    opt_con_req_dict = {"-c": ["-m"], "-s": ["-t"], "-S": ["-f", "-k"]}
+    opt_con_req_dict = {"-c": ["-m"], "-s": ["-t"]}
     opt_multi_list = ["-f", "-s", "-t", "-S"]
     opt_val_list = ["-i", "-m", "-o", "-s", "-t", "-y", "-F", "-S", "-k"]
     opt_valid_val = {"-k": ["and", "or"]}